optimisation

+++ benches/my_benchmark.rs
@@ -14,7 +14,7 @@ pub fn criterion_benchmark(c: &mut Criterion) {
 
             for x in 0..7000 {
                 for y in 0..3000 {
-                    let starship = Starship::new(x as f32, y as f32, 0, 0, 0, 0., 0.);
+                    let starship = Starship::new(x, y, 0, 0, 0, 0., 0.);
                     black_box(game.starship_is_crash(starship));
                 }
             }
@@ -24,7 +24,7 @@ pub fn criterion_benchmark(c: &mut Criterion) {
         b.iter(|| {
             for rotation in -90..=90 {
                 for power in 0..=4 {
-                    let mut starship = Starship::new(0., 0., 10000, rotation, power, 0., 0.);
+                    let mut starship = Starship::new(0, 0, 10000, rotation, power, 0., 0.);
                     for _ in 0..50 {
                         black_box(starship.apply_movement());
                     }
